+++ armsrc/mifarecmd.h
@@ -60,6 +60,7 @@ void MifareGen3Freez(void); // Gen 3 magic card lock further UID changes
 
 // MFC GEN4 GDM
 void MifareReadConfigBlockGDM(uint8_t *key);
+void MifareWriteConfigBlockGDM(uint8_t *datain);
 void MifareWriteBlockGDM(uint8_t blockno, uint8_t keytype, uint8_t *key, uint8_t *datain);
 
 // MFC GEN4 GTU